Update Shifting array
A left rotation operation on an array shifts each of the array's elements 1 unit to the left. 
Given an array a of n integers and a number,d ,perform d left rotations on the array. 
Return the updated array to be printed as a single line of space-separated integers.

Function Description

Complete the function rotLeft in the editor below. It should return the resulting array of integers.

rotLeft has the following parameter(s):

An array of integers a.
An integer d, the number of rotations.